pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["KeitoAV <keito_ekka@mail.ru>"]
readme = "README.md"
packages = [{include = "test_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
drf-spectacular = "^0.25.1"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
